Extract flash-related methods to their own module
