+++ app/controllers/application_controller.rb
@@ -5,6 +5,7 @@ class ApplicationController < ActionController::Base
 
   include Turbo::Redirection
   include Pagy::Backend
+  include Flash
 
   before_action :store_user_location!, if: :storable_location?
   # The callback which stores the current location must be added before you
@@ -128,33 +129,6 @@ class ApplicationController < ActionController::Base
     # https://stackoverflow.com/a/69313330/
     def set_current_user
       Current.user = current_user
-    end
-
-    # Renders a flash message via Turbo Stream.
-    #
-    # Usage:
-    # You can use :notice, :success, :alert and :error, see flash/_message.html.erb
-    # > flash.now[:success] = "Profile updated"
-    # > render_flash
-    def render_flash
-      return if flash.empty?
-
-      render turbo_stream: turbo_stream.prepend("flash-messages", partial: "flash/message")
-    end
-
-    # Renders a flash success message for turbo_stream and html formats.
-    # Usage: respond_with_flash_success(I18n.t("feedback.success"))
-    def respond_with_flash_success(message)
-      respond_to do |format|
-        format.turbo_stream do
-          flash.now[:success] = message
-          render_flash
-        end
-        format.html do
-          flash.keep[:success] = message
-          redirect_back(fallback_location: fallback_location)
-        end
-      end
     end
 
     # Ensures that the current request is a Turbo Frame request.
